json: move all implemention methods of Streamer to .h
