+++ source/common/json/BUILD
@@ -48,7 +48,6 @@ envoy_cc_library(
 
 envoy_cc_library(
     name = "json_streamer_lib",
-    srcs = ["json_streamer.cc"],
     hdrs = ["json_streamer.h"],
     deps = [
         ":json_sanitizer_lib",
